requirements.txt
Flask
psycopg2-binary
pandas
requests
python-dotenv
sqlalchemy